# server/swagger.yaml
swagger: '2.0'
info:
  description: Mergin API to synchronize your GIS data.
  version: '0.6'
  title: Mergin API
basePath: /v1
...
schemes:
  - https
  - http
definitions:
...
  ProjectListItem:
    allOf:
      - $ref: '#/definitions/Project'
      - type: object
        properties:
          name:
            type: string
            example: mergin
          namespace:
            type: string
            example: mergin
          creator:
            type: integer
            example: 1
            description: project creator ID
          disk_usage:
            type: integer
            example: 25324373
            description: project size in bytes
          permissions:
            type: object
            properties:
              delete:
                type: boolean
                example: false
              update:
                type: boolean
                example: false
              upload:
                type: boolean
                example: true
          tags:
            type: array
            items:
              $ref: '#/definitions/MerginTag'
          updated:
            type: string
            format: date-time
            example: '2018-11-30T08:47:58.636074Z'
            description: last modified
          version:
            type: string
            example: 'v2'
            description: last project version
          uploads:
            type: array
            items:
              type: string
              example: "669b838e-a30b-4338-b2b6-3da144742a82"
              description: uuid for ongoing upload
          access:
            $ref: '#/definitions/Access'
  ProjectDetail:
    allOf:
      - $ref: '#/definitions/ProjectListItem'
      - type: object
        properties:
          files:
            type: array
            items:
              allOf:
                - $ref: '#/definitions/FileInfo'
                - $ref: '#/definitions/HistoryFileInfo'
...
paths:
...
  /project:
    get:
      tags:
        - project
      summary: List mergin projects.
      description: Returns limited list of projects, optionally filtered by tags, search query, username.
      operationId: get_projects
      produces:
        - application/json
      parameters:
        - name: tags
          in: query
          description: Filter by mergin tags.
          required: false
          type: array
          items:
            type: string
            enum: *MERGIN_TAG
        - name: q
          in: query
          description: Search query string.
          required: false
          type: string
        - name: user
          in: query
          description: Username for 'flag' filter. If not provided, it means user executing request.
          required: false
          type: string
        - name: flag
          in: query
          description: Predefined filter flag.
          required: false
          type: string
          enum:
            - created
            - shared
        - name: limit
          in: query
          description: Maximum number of returned projects.
          required: false
          type: integer
          minimum: 1
      responses:
        '200':
          description: Success.
          schema:
            type: array
            items:
              $ref: '#/definitions/ProjectListItem'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
  '/project/{namespace}':
    post:
      tags:
        - project
      summary: Add a new mergin project.
      description: Add new project to database and create empty project directory. Project name is project dir at the same time.
      operationId: add_project
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
        - in: body
          name: project
          description: Project object that needs to be added to the database.
          required: true
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
                example: mergin
              public:
                type: boolean
                example: true
              template:
                type: string
                example: Template
      responses:
        '200':
          $ref: '#/responses/Success'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '405':
          description: Invalid input
        '422':
          $ref: '#/responses/InvalidDataResp'
      x-swagger-router-controller: src.controllers.project_controller
  '/project/{namespace}/{project_name}':
    get:
      tags:
        - project
      summary: Find project by name.
      description: Returns a single project of specified version with details about files including history for versioned files (diffs) if needed.
      operationId: get_project
      produces:
        - application/json
      parameters:
        - name: project_name
          in: path
          description: Name of project to return.
          required: true
          type: string
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
        - name: since
          in: query
          description: Version to look up diff files history from.
          required: false
          type: string
        - name: version
          in: query
          description: Project version.
          required: false
          type: string
      responses:
        '200':
          description: Success.
          schema:
            $ref: '#/definitions/ProjectDetail'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
    put:
     tags:
       - project
     summary: Update an existing project.
     description: Updates 'public' flag and access list for project.
     operationId: update_project
     consumes:
       - application/json
     produces:
       - application/json
     parameters:
       - name: project_name
         in: path
         description: Name of project that need to be updated.
         required: true
         type: string
       - name: namespace
         in: path
         description: Namespace for project to look into.
         required: true
         type: string
       - in: body
         name: data
         description: Data to be updated.
         required: true
         schema:
           type: object
           properties:
             access:
               type: object
               $ref: '#/definitions/Access'
     responses:
       '200':
         $ref: '#/responses/Success'
       '400':
         $ref: '#/responses/BadStatusResp'
       '403':
         $ref: '#/responses/Forbidden'
       '404':
         $ref: '#/responses/NotFoundResp'
       '405':
         description: Validation exception.
     x-swagger-router-controller: src.controllers.project_controller
    delete:
      tags:
        - project
      summary: Delete a project.
      description: Remove project from database and project directory with all files.
      operationId: delete_project
      produces:
        - application/json
      parameters:
        - name: project_name
          in: path
          description: Name of project to delete.
          required: true
          type: string
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
      responses:
        '200':
          $ref: '#/responses/Success'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
  '/project/by_names':
    post:
      tags:
        - project
      summary: Find projects by name.
      description: Returns list of requested projects
      operationId: get_projects_by_names
      produces:
        - application/json
      parameters:
        - in: body
          name: data
          description: List of requested projects.
          required: false
          schema:
            type: object
            properties:
              projects:
                type: array
                items:
                  type: string
            example:
              projects: [mergin_namespace/project_name]
      responses:
        '200':
          description: Success.
          schema:
            type: object
            additionalProperties:
                type: object
                $ref: '#/definitions/ProjectListItem'
        '400':
          $ref: '#/responses/BadStatusResp'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
...
  '/project/version/{namespace}/{project_name}':
    get:
      tags:
        - project
      summary: Get versions (history) of project.
      description: Returns a list of project versions with changes information.
      operationId: get_project_versions
      produces:
        - application/json
      parameters:
        - name: project_name
          in: path
          description: Name of project to return versions for.
          required: true
          type: string
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
        - name: version_id
          in: query
          required: false
          type: string
      responses:
        '200':
          description: Success.
          schema:
            type: array
            items:
              $ref: '#/definitions/ProjectVersion'
        '400':
          $ref: '#/responses/BadStatusResp'
        '403':
          $ref: '#/responses/Forbidden'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
  '/project/download/{namespace}/{project_name}':
    get:
      tags:
        - project
      summary: Download dir for single project.
      description: Download whole project folder as zip file or produces stream to parse.
      operationId: download_project
      produces:
        - application/zip
        - multipart/form-data
      parameters:
        - name: project_name
          in: path
          description: Name of project to download.
          required: true
          type: string
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
        - name: format
          in: query
          description: Output format (only zip available).
          required: false
          type: string
          enum:
            - zip
        - name: version
          in: query
          description: Version tag.
          required: false
          type: string
      responses:
        '200':
          $ref: '#/responses/Success'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
        '422':
          $ref: '#/responses/InvalidDataResp'
      x-swagger-router-controller: src.controllers.project_controller
  '/project/raw/{namespace}/{project_name}':
    get:
      tags:
        - project
      summary: Download project file
      description: Download individual file or it's diff file from project.
      operationId: download_project_file
      produces:
        - application/octet-stream
      parameters:
        - name: project_name
          in: path
          description: Project name.
          required: true
          type: string
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
        - name: file
          in: query
          description: Path to file.
          required: true
          type: string
        - name: version
          in: query
          description: Version tag.
          required: false
          type: string
        - name: diff
          in: query
          description: Ask for diff file instead of full one.
          required: false
          type: boolean
      responses:
        '200':
          $ref: '#/responses/Success'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
        '422':
          $ref: '#/responses/InvalidDataResp'
      x-swagger-router-controller: src.controllers.project_controller
  /project/push/{namespace}/{project_name}:
    post:
      tags:
       - project
      summary: Synchronize project data.
      description: Apply changes in project if no uploads required. Creates upload transaction for added/modified files.
      operationId: project_push
      consumes:
       - application/json
      produces:
       - application/json
      parameters:
        - name: namespace
          in: path
          description: Namespace for project to look into.
          required: true
          type: string
        - name: project_name
          in: path
          description: Project name.
          required: true
          type: string
        - in: body
          name: data
          required: true
          description: Description of project changes.
          schema:
            type: object
            required:
              - version
              - changes
            properties:
              version:
                type: string
                example: v1
              changes:
                type: object
                required:
                  - added
                  - updated
                  - renamed
                  - removed
                properties:
                  added:
                    type: array
                    items:
                      $ref: '#/definitions/UploadFileInfo'
                  updated:
                    type: array
                    items:
                      $ref: '#/definitions/UpdateFileInfo'
                  renamed:
                    type: array
                    items:
                      $ref: '#/definitions/FileInfo'
                  removed:
                    type: array
                    items:
                      $ref: '#/definitions/FileInfo'
      responses:
        '200':
          $ref: '#/responses/Success'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
...
  /project/push/finish/{transaction_id}:
    post:
      tags:
       - project
      summary: Finalize project data upload.
      operationId: push_finish
      parameters:
        - name: transaction_id
          in: path
          description: Transaction id.
          required: true
          type: string
      responses:
        '200':
          $ref: '#/responses/Success'
        '400':
          $ref: '#/responses/BadStatusResp'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/Forbidden'
        '404':
          $ref: '#/responses/NotFoundResp'
      x-swagger-router-controller: src.controllers.project_controller
...
responses:
  UnauthorizedError:
    description: Authentication information is missing or invalid.
    headers:
      WWW_Authenticate:
        type: string
  NotFoundResp:
    description: Project not found.
  BadStatusResp:
    description: Invalid request.
  InvalidDataResp:
    description: Invalid/unprocessable data.
  Success:
    description: Success.
  Forbidden:
    description: Access is denied.
...